storybook上で正しくメッセージリソースが扱えるように修正
ns=["common"] が悪さをしていたっぽい
日英切り替えはできていないがとりあえずよしとする
