+++ src/@types/react-i18next.d.ts
@@ -1,13 +1,13 @@
 import 'react-i18next';
-import ns1 from '../locale/resources_ja.json';
-import ns2 from '../locale/resources_en.json';
+import { resourceJap } from '../locale/resources_ja';
+import { resourceEng } from '../locale/resources_en';
 
 declare module 'react-i18next' {
   interface CustomTypeOptions {
     defaultNS: 'ns1';
     resources: {
-      ns1: typeof ns1;
-      ns2: typeof ns2;
+      ns1: typeof resourceJap;
+      ns2: typeof resourceEng;
     };
   }
 }
