+++ src/i18n.ts
@@ -1,16 +1,14 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
-import resourcesJa from './locale/resources_ja.json';
-import resourcesEn from './locale/resources_en.json';
+import { resourceJap } from './locale/resources_ja';
+import { resourceEng } from './locale/resources_en';
 
 const resources = {
-  en: {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    translation: resourcesEn,
-  },
   ja: {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    translation: resourcesJa,
+    translation: resourceJap,
+  },
+  en: {
+    translation: resourceEng,
   },
 };
